pyproject.toml
[tool.poetry]
name = "email-clean"
version = "0.1.0"
description = ""
authors = ["Mitch Hung <mitch.hung@gro-intelligence.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.80.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
